+++ packages/terra-framework-docs/CHANGELOG.md
@@ -1,8 +1,15 @@
 # Changelog
 
+## Unreleased
+
+* Changed
+  * Migrated terra-collapsible-menu-view-docs
+  * Migrated docs from date-picker.
+  * Migrated docs from date-time-picker
+
 * Added
-  * Migrate docs from terra-menu
   * Migrate docs from terra-navigation-layout
+  * Migrate docs from terra-menu
   * Migrate docs from terra-infinite-list
 
 ## 1.7.0 - (March 8, 2022)